backend/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
sqlalchemy==1.4.47  # Use a compatible version for databases
passlib[bcrypt]==1.7.4
python-jose==3.3.0
pydantic==1.10.9
databases==0.6.0

